Simplify abus_current_url
I'm having trouble with this function on pantheon.io. The reason is that the $_SERVER[ 'SERVER_PORT' ] variable is actually 9172 or something and the site still runs on 80 though. I'm not sure exactly why that is the case but I assume it has to do with the fact that everything is running in docker containers which each have their own port.

I've ran some experiments with the code above, even on a local MAMP running on :8000 and it's working just fine with this code update. 

This code update would also honour whatever setting was done within WordPress in terms of which protocol should be used etc.

Am I missing anything?